requirements.txt
django
django-cors-headers
django-mailer
djangorestframework
psycopg2-binary
gunicorn
celery
dj-rest-auth
django-allauth
flower
redis
requests
django-celery-results